move providers code to providers.rs
