+++ src/providers.rs
@@ -1,11 +1,61 @@
-use crate::AppState;
+use crate::{
+    auth::{
+        ecmwfapi_provider::{EcmwfApiProvider, EcmwfApiProviderConfig},
+        jwt_provider::{JWTAuthConfig, JWTProvider},
+        openid_offline_provider::{OpenIDOfflineProvider, OpenIDOfflineProviderConfig},
+        plain_provider::{PlainAuthConfig, PlainAuthProvider},
+    },
+    models::User,
+    AppState,
+};
 use axum::{
     extract::{ConnectInfo, State},
     Json,
 };
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use std::net::SocketAddr;
 use tracing::{debug, info};
+
+/// Configuration options for each authentication provider.
+#[derive(Deserialize, Serialize, JsonSchema, Debug)]
+#[serde(tag = "type")]
+pub enum ProviderConfig {
+    #[serde(rename = "ecmwf-api")]
+    EcmwfApiAuthConfig(EcmwfApiProviderConfig),
+
+    #[serde(rename = "jwt")]
+    JWTAuthConfig(JWTAuthConfig),
+
+    #[serde(rename = "openid-offline")]
+    OpenIDOfflineAuthConfig(OpenIDOfflineProviderConfig),
+
+    #[serde(rename = "plain")]
+    PlainAuthConfig(PlainAuthConfig),
+}
+
+/// An authentication provider must be able to return a User or an error.
+#[async_trait::async_trait]
+pub trait Provider: Send + Sync {
+    fn get_name(&self) -> &str;
+    fn get_type(&self) -> &str;
+    /// Providers that support realm-based filtering should override this.
+    fn get_realm(&self) -> Option<&str> {
+        None
+    }
+    async fn authenticate(&self, credentials: &str) -> Result<User, String>;
+}
+
+/// Create an authentication provider from a given config.
+pub fn create_auth_provider(config: &ProviderConfig) -> Box<dyn Provider> {
+    match config {
+        ProviderConfig::EcmwfApiAuthConfig(cfg) => Box::new(EcmwfApiProvider::new(cfg)),
+        ProviderConfig::JWTAuthConfig(cfg) => Box::new(JWTProvider::new(cfg)),
+        ProviderConfig::OpenIDOfflineAuthConfig(cfg) => Box::new(OpenIDOfflineProvider::new(cfg)),
+        ProviderConfig::PlainAuthConfig(cfg) => Box::new(PlainAuthProvider::new(cfg)),
+    }
+}
 
 /// GET /providers endpoint: returns only the provider "name", "type" and "realm" fields
 /// by converting each provider config into JSON and extracting the fields.
